+++ src/main/java/com/l2jserver/gameserver/ai/L2AttackableAI.java
@@ -534,7 +534,7 @@ public class L2AttackableAI extends L2CharacterAI implements Runnable
 					}
 					
 					// Get the hate level of the L2Attackable against this L2Character target contained in _aggroList
-					int hating = npc.getHating(target);
+					long hating = npc.getHating(target);
 					
 					// Add the attacker to the L2Attackable _aggroList with 0 damage and 1 hate
 					if (hating == 0)
@@ -559,8 +559,7 @@ public class L2AttackableAI extends L2CharacterAI implements Runnable
 			if ((hated != null) && !npc.isCoreAIDisabled())
 			{
 				// Get the hate level of the L2Attackable against this L2Character target contained in _aggroList
-				int aggro = npc.getHating(hated);
-				
+				long aggro = npc.getHating(hated);
 				if ((aggro + _globalAggro) > 0)
 				{
 					// Set the L2Character movement type to run and send Server->Client packet ChangeMoveType to all others L2PcInstance
@@ -2494,7 +2493,7 @@ public class L2AttackableAI extends L2CharacterAI implements Runnable
 	 * @param aggro The value of hate to add to the actor against the target
 	 */
 	@Override
-	protected void onEvtAggression(L2Character target, int aggro)
+	protected void onEvtAggression(L2Character target, long aggro)
 	{
 		final L2Attackable me = getActiveChar();
 		if (me.isDead())
