+++ src/main/java/com/l2jserver/gameserver/ai/L2DoorAI.java
@@ -93,7 +93,7 @@ public class L2DoorAI extends L2CharacterAI
 	}
 	
 	@Override
-	protected void onEvtAggression(L2Character target, int aggro)
+	protected void onEvtAggression(L2Character target, long aggro)
 	{
 	}
