+++ src/main/java/com/l2jserver/gameserver/ai/L2SiegeGuardAI.java
@@ -282,7 +282,7 @@ public class L2SiegeGuardAI extends L2CharacterAI implements Runnable
 				if (autoAttackCondition(target)) // check aggression
 				{
 					// Get the hate level of the L2Attackable against this L2Character target contained in _aggroList
-					int hating = npc.getHating(target);
+					long hating = npc.getHating(target);
 					
 					// Add the attacker to the L2Attackable _aggroList with 0 damage and 1 hate
 					if (hating == 0)
@@ -308,7 +308,7 @@ public class L2SiegeGuardAI extends L2CharacterAI implements Runnable
 			if (hated != null)
 			{
 				// Get the hate level of the L2Attackable against this L2Character target contained in _aggroList
-				int aggro = npc.getHating(hated);
+				long aggro = npc.getHating(hated);
 				
 				if ((aggro + _globalAggro) > 0)
 				{
@@ -835,7 +835,7 @@ public class L2SiegeGuardAI extends L2CharacterAI implements Runnable
 	 * @param aggro The value of hate to add to the actor against the target
 	 */
 	@Override
-	protected void onEvtAggression(L2Character target, int aggro)
+	protected void onEvtAggression(L2Character target, long aggro)
 	{
 		if (_actor == null)
 		{
