+++ src/main/java/com/l2jserver/gameserver/model/actor/instance/L2FortCommanderInstance.java
@@ -73,7 +73,7 @@ public class L2FortCommanderInstance extends L2DefenderInstance
 	}
 	
 	@Override
-	public void addDamageHate(L2Character attacker, int damage, int aggro)
+	public void addDamageHate(L2Character attacker, int damage, long aggro)
 	{
 		if (attacker == null)
 		{
