+++ src/main/java/com/l2jserver/gameserver/model/events/AbstractScript.java
@@ -3221,7 +3221,7 @@ public abstract class AbstractScript implements INamable
 	 * @param skill the skill to cast
 	 * @param desire the desire to cast the skill
 	 */
-	protected void addSkillCastDesire(L2Npc npc, L2Character target, SkillHolder skill, int desire)
+	protected void addSkillCastDesire(L2Npc npc, L2Character target, SkillHolder skill, long desire)
 	{
 		addSkillCastDesire(npc, target, skill.getSkill(), desire);
 	}
@@ -3233,7 +3233,7 @@ public abstract class AbstractScript implements INamable
 	 * @param skill the skill to cast
 	 * @param desire the desire to cast the skill
 	 */
-	protected void addSkillCastDesire(L2Npc npc, L2Character target, Skill skill, int desire)
+	protected void addSkillCastDesire(L2Npc npc, L2Character target, Skill skill, long desire)
 	{
 		if (npc instanceof L2Attackable)
 		{
